game splitted into modules
