+++ constants.py
@@ -1,4 +1,2 @@
 BOARD_SIZE = 4
 EMPTY_CELL = ' '
-
-
