+++ custom_types.py
@@ -3,4 +3,3 @@ from typing import NamedTuple
 class Move(NamedTuple):
     row: int
     col: int
-
